fix issue scrutinizers

+++ component/admin/controllers/order.php
@@ -553,11 +553,11 @@ class RedshopControllerOrder extends RedshopController
 
 			$days = (date("H") > $clock && $days == 0) ? 1 : $days;
 
-			$product_download_days_time = (int) (time() + ($days * 24 * 60 * 60));
+			$product_download_days_time = (time() + ($days * 24 * 60 * 60));
 
 			$endtime = mktime(
-				$clock, $clock_min, 0, date("m", $product_download_days_time), date("d", $product_download_days_time),
-				date("Y", $product_download_days_time)
+				$clock, $clock_min, 0, date("m", (int) $product_download_days_time), date("d", (int) $product_download_days_time),
+				date("Y", (int) $product_download_days_time)
 			);
 
 			// If download product is set to infinit
